+++ changelog.rst
@@ -8,6 +8,7 @@ ChangeLog
 - fix wrong name from obj2driver.
 - fix copy of function.
 - enhance use of params.
+- fix updatename.
 
 0.0.3 - 2016/10/26
 ==================